requirements.txt
flask
faiss-cpu
pandas
openai
sentence-transformers
